+++ MMU.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <stdint.h>
 
-uint8_t readByte(int address);
-uint16_t readShort(int address);
+uint8_t readByte(uint16_t address);
+uint16_t readShort(uint16_t address);
 void loadCartridge(char* path);
 void initMMU();
